+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/ComposeSearchTest.kt
@@ -156,7 +156,7 @@ class ComposeSearchTest {
 
     @SmokeTest
     @Test
-    fun scanButtonDenyPermissionTest() {
+    fun verifyQRScanningCameraAccessDialogTest() {
         val cameraManager = TestHelper.appContext.getSystemService(Context.CAMERA_SERVICE) as CameraManager
         Assume.assumeTrue(cameraManager.cameraIdList.isNotEmpty())
 
@@ -177,7 +177,7 @@ class ComposeSearchTest {
 
     @SmokeTest
     @Test
-    fun scanButtonAllowPermissionTest() {
+    fun scanQRCodeToOpenAWebpageTest() {
         val cameraManager = TestHelper.appContext.getSystemService(Context.CAMERA_SERVICE) as CameraManager
         Assume.assumeTrue(cameraManager.cameraIdList.isNotEmpty())
 
@@ -213,7 +213,7 @@ class ComposeSearchTest {
     // Verifies a temporary change of search engine from the Search shortcut menu
     @SmokeTest
     @Test
-    fun selectSearchEnginesShortcutTest() {
+    fun searchEnginesCanBeChangedTemporarilyFromSearchSelectorMenuTest() {
         val enginesList = listOf("DuckDuckGo", "Google", "Amazon.com", "Wikipedia", "Bing", "eBay")
 
         enginesList.forEach {
@@ -230,7 +230,7 @@ class ComposeSearchTest {
     }
 
     @Test
-    fun accessSearchSettingFromSearchSelectorMenuTest() {
+    fun defaultSearchEnginesCanBeSetFromSearchSelectorMenuTest() {
         searchScreen {
             clickSearchSelectorButton()
         }.clickSearchEngineSettings {
@@ -247,7 +247,7 @@ class ComposeSearchTest {
     }
 
     @Test
-    fun clearSearchTest() {
+    fun verifyClearSearchButtonTest() {
         homeScreen {
         }.openSearch {
             typeSearch(queryString)
@@ -622,9 +622,10 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/1850517
     // Test that verifies the Firefox Suggest results in a general search context
     @Test
-    fun firefoxSuggestHeaderForBrowsingDataSuggestionsTest() {
+    fun verifyFirefoxSuggestHeaderForBrowsingDataSuggestionsTest() {
         val firstPage = TestAssetHelper.getGenericAsset(searchMockServer, 1)
         val secondPage = TestAssetHelper.getGenericAsset(searchMockServer, 2)
 
@@ -646,8 +647,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154197
     @Test
-    fun verifySearchTabsItemsTest() {
+    fun verifyTabsSearchItemsTest() {
         navigationToolbar {
         }.clickUrlbar {
             clickSearchSelectorButton()
@@ -659,8 +661,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154198
     @Test
-    fun verifySearchTabsWithoutOpenTabsTest() {
+    fun verifyTabsSearchWithoutOpenTabsTest() {
         navigationToolbar {
         }.clickUrlbar {
             clickSearchSelectorButton()
@@ -672,9 +675,10 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154199
     @SmokeTest
     @Test
-    fun verifySearchTabsWithOpenTabsTest() {
+    fun verifyTabsSearchWithOpenTabsTest() {
         val firstPageUrl = TestAssetHelper.getGenericAsset(searchMockServer, 1)
         val secondPageUrl = TestAssetHelper.getGenericAsset(searchMockServer, 2)
 
@@ -707,8 +711,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154203
     @Test
-    fun verifySearchForBookmarksUITest() {
+    fun verifyBookmarksSearchItemsTest() {
         navigationToolbar {
         }.clickSearchSelectorButton {
             selectTemporarySearchMethod("Bookmarks")
@@ -719,8 +724,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154204
     @Test
-    fun bookmarkSearchWithNoBookmarksTest() {
+    fun verifyBookmarkSearchWithNoBookmarksTest() {
         navigationToolbar {
         }.clickSearchSelectorButton {
             selectTemporarySearchMethod("Bookmarks")
@@ -729,8 +735,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154206
     @Test
-    fun bookmarkSearchWhenBookmarksExistTest() {
+    fun verifyBookmarksSearchForBookmarkedItemsTest() {
         createBookmarkItem(url = "https://bookmarktest1.com", title = "Test1", position = 1u)
         createBookmarkItem(url = "https://bookmarktest2.com", title = "Test2", position = 2u)
 
@@ -756,8 +763,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154212
     @Test
-    fun verifySearchHistoryItemsTest() {
+    fun verifyHistorySearchItemsTest() {
         navigationToolbar {
         }.clickUrlbar {
             clickSearchSelectorButton()
@@ -769,8 +777,9 @@ class ComposeSearchTest {
         }
     }
 
+    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/2154213
     @Test
-    fun verifySearchHistoryWithoutBrowsingDataTest() {
+    fun verifyHistorySearchWithoutBrowsingHistoryTest() {
         navigationToolbar {
         }.clickUrlbar {
             clickSearchSelectorButton()
